feat: 주식 전체 종목 조회

+++ BE/src/main/java/com/example/LAGO/repository/StockInfoRepository.java
@@ -2,6 +2,6 @@ package com.example.LAGO.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import com.example.LAGO.domain.StockInfo;
-public interface StockInfoRepository extends JpaRepository<StockInfo, Long> {
+public interface StockInfoRepository extends JpaRepository<StockInfo, Integer> {
 
 }
